requirements.txt
aiomysql==0.2.0
aiosqlite==0.21.0
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
asyncpg==0.30.0
certifi==2025.1.31
click==8.1.8
coverage==7.8.0
Deprecated==1.2.18
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
limits==5.0.0
Mako==1.3.10
MarkupSafe==3.0.2
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
psycopg2==2.9.10
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
PyJWT==2.10.1
PyMySQL==1.1.1
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-cov==6.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlmodel==0.0.24
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzlocal==5.3.1
uvicorn==0.34.2
wrapt==1.17.2
